Always convert JavaScript number value to Java double value
422768: [BrowserFunction] JavaScript Number must always be converted to
Java Double
https://bugs.eclipse.org/bugs/show_bug.cgi?id=422768

+++ tests/org.eclipse.rap.rwt.test/src/org/eclipse/rap/rwt/widgets/BrowserUtil_Test.java
@@ -117,7 +117,7 @@ public class BrowserUtil_Test {
     Fixture.fakeCallOperation( getId( browser ), "evaluationSucceeded", parameters );
     Fixture.readDataAndProcessAction( browser );
 
-    verify( browserCallback ).evaluationSucceeded( Integer.valueOf( 5 ) );
+    verify( browserCallback ).evaluationSucceeded( Double.valueOf( 5 ) );
   }
 
   @Test
